+++ src/main/kotlin/net/ndrei/teslacorelib/gui/RedstoneTogglePiece.kt
@@ -4,7 +4,7 @@ import net.minecraft.client.renderer.GlStateManager
 import net.ndrei.teslacorelib.MOD_ID
 import net.ndrei.teslacorelib.inventory.BoundingRectangle
 import net.ndrei.teslacorelib.inventory.IRedstoneControlledMachine
-import net.ndrei.teslacorelib.localization.GuiPieceType
+import net.ndrei.teslacorelib.localization.GUI_REDSTONE
 import net.ndrei.teslacorelib.localization.localizeModString
 
 /**
@@ -35,9 +35,9 @@ class RedstoneTogglePiece(private val machine: IRedstoneControlledMachine, left:
 
     override fun getStateToolTip(state: Int): List<String> =
         when (state % 3) {
-            0 -> listOf(localizeModString(MOD_ID, GuiPieceType.REDSTONE.key, "Always Active").formattedText)
-            1 -> listOf(localizeModString(MOD_ID, GuiPieceType.REDSTONE.key, "Active on Redstone Signal").formattedText)
-            2 -> listOf(localizeModString(MOD_ID, GuiPieceType.REDSTONE.key, "Active without Redstone Signal").formattedText)
+            0 -> listOf(localizeModString(MOD_ID, GUI_REDSTONE, "Always Active").formattedText)
+            1 -> listOf(localizeModString(MOD_ID, GUI_REDSTONE, "Active on Redstone Signal").formattedText)
+            2 -> listOf(localizeModString(MOD_ID, GUI_REDSTONE, "Active without Redstone Signal").formattedText)
             else -> listOf()
         }