+++ src/main/kotlin/net/ndrei/teslacorelib/localization/GuiPieceType.kt
@@ -1,8 +1,8 @@
 package net.ndrei.teslacorelib.localization
 
-enum class GuiPieceType(val key: String) {
-    SIDE_CONFIG("side config"),
-    REDSTONE("redstone"),
-    FLUID_TANK("fluid tank"),
-    LOCKABLE_INVENTORY("lockable inventory")
-}
+const val GUI_SIDE_CONFIG = "side config"
+const val GUI_REDSTONE = "redstone"
+const val GUI_FLUID_TANK = "fluid tank"
+const val GUI_LOCKABLE_INVENTORY = "lockable inventory"
+const val GUI_ENERGY = "energy"
+const val GUI_WORK_ENERGY = "work energy"